+++ resources/[dg]/dg-minigames/typescript/src/client/client.ts
@@ -1,3 +1,4 @@
 import './modules/keygame';
 import './modules/gridgames';
+import './modules/keypad';
 import './controllers/events';