+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First Person Cam die gebruikt werd bij pdcam en verrekijker verplaatst naar eigen module om makkelijker opnieuw te gebruiken
 - Je kan nu rechtstreeks fotos maken met de PD Camera
 - Nieuws gerelateerde items toegevoegd
+- Keypad UI toegevoegd
 
 ## Changed
 - Veel item decaysnelheiden verlaagd